+++ src/main/java/com/order/orderlink/ai/exception/AiApiException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.ai.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class AiApiException extends BaseException {
