+++ src/main/java/com/order/orderlink/ai/presentation/AiApiController.java
@@ -15,7 +15,7 @@ import com.order.orderlink.ai.application.dtos.AiApiResponse;
 import com.order.orderlink.common.auth.UserDetailsImpl;
 import com.order.orderlink.common.dtos.SuccessResponse;
 import com.order.orderlink.common.enums.SuccessCode;
-import com.order.orderlink.common.exception.AiApiException;
+import com.order.orderlink.ai.exception.AiApiException;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
