+++ src/main/java/com/order/orderlink/common/client/FoodClient.java
@@ -3,8 +3,6 @@ package com.order.orderlink.common.client;
 import org.springframework.stereotype.Component;
 import org.springframework.web.reactive.function.client.WebClient;
 
-import com.order.orderlink.common.dtos.FoodRequest;
-
 import lombok.RequiredArgsConstructor;
 
 @Component
@@ -21,14 +19,4 @@ public class FoodClient {
 			.bodyToMono(String.class)
 			.block(); // 동기 호출
 	}
-
-	public String createFood(FoodRequest.Create request) {
-		return webClient.post()
-			.uri(FOODS_URI)
-			.bodyValue(request)
-			.retrieve()
-			.bodyToMono(String.class)
-			.block(); // 동기 호출
-	}
-
 }
