+++ src/main/java/com/order/orderlink/food/exception/FoodException.java
@@ -1,10 +1,11 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.food.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class FoodException extends BaseException {
 
-    public FoodException(ErrorCode errorCode) {
-        super(errorCode.getStatus(), errorCode.getMessage());
-    }
+	public FoodException(ErrorCode errorCode) {
+		super(errorCode.getStatus(), errorCode.getMessage());
+	}
 }
