+++ src/main/java/com/order/orderlink/order/exception/OrderException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.order.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class OrderException extends BaseException {
 	public OrderException(ErrorCode errorCode) {
