+++ src/main/java/com/order/orderlink/payment/exception/PaymentException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.payment.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class PaymentException extends BaseException {
 	public PaymentException(ErrorCode errorCode) {
