+++ src/main/java/com/order/orderlink/region/exception/RegionException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.region.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class RegionException extends BaseException {
 	public RegionException(ErrorCode errorCode) {
