+++ src/main/java/com/order/orderlink/restaurant/exception/RestaurantException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.restaurant.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class RestaurantException extends BaseException {
 	public RestaurantException(ErrorCode errorCode) {
