+++ src/main/java/com/order/orderlink/review/application/ReviewService.java
@@ -13,9 +13,9 @@ import com.order.orderlink.common.client.OrderClient;
 import com.order.orderlink.common.client.RestaurantClient;
 import com.order.orderlink.common.client.UserClient;
 import com.order.orderlink.common.enums.ErrorCode;
-import com.order.orderlink.common.exception.OrderException;
-import com.order.orderlink.common.exception.RestaurantException;
-import com.order.orderlink.common.exception.ReviewException;
+import com.order.orderlink.order.exception.OrderException;
+import com.order.orderlink.restaurant.exception.RestaurantException;
+import com.order.orderlink.review.exception.ReviewException;
 import com.order.orderlink.order.domain.Order;
 import com.order.orderlink.orderitem.domain.OrderItem;
 import com.order.orderlink.restaurant.domain.Restaurant;
