+++ src/main/java/com/order/orderlink/review/exception/ReviewException.java
@@ -1,6 +1,7 @@
-package com.order.orderlink.common.exception;
+package com.order.orderlink.review.exception;
 
 import com.order.orderlink.common.enums.ErrorCode;
+import com.order.orderlink.common.exception.BaseException;
 
 public class ReviewException extends BaseException {
